adding mainwindow (Docs.txt is excluded)

+++ PipeEngine.cs
@@ -7,6 +7,7 @@ using System.Threading;
 using System.IO.Pipes;
 using System.Diagnostics;
 using System.Timers;
+using System.IO;
 
 namespace ProjectA
 {
@@ -20,7 +21,7 @@ namespace ProjectA
         public byte[] result;
         public int resultLength;
 
-        public PipeEngine(string PipeName,string FileName = "python\\pipe.py",int IOBuffSize = 65535,int timeout = 5000)
+        public PipeEngine(string PipeName,string FileName = "python\\pipe.py",int IOBuffSize = 65535,int timeout = 1000)
         {
             this.timeout = timeout;
             this.PipeName = PipeName;
@@ -33,8 +34,26 @@ namespace ProjectA
             P.StartInfo.UseShellExecute = true;    //是否使用操作系统shell启动
             P.StartInfo.CreateNoWindow = true;  //不显示程序窗口
             P.StartInfo.WindowStyle = System.Diagnostics.ProcessWindowStyle.Minimized;
+            CheckPythonEnviroment();
         }
 
+        private static void CheckPythonEnviroment()
+        {
+            if(!System.IO.File.Exists("PythonHasInstalled"))
+            try
+            {
+                Process P1 = Process.Start("python.exe","python\\pakcheck.py");
+                FileStream fs1 = new FileStream("PythonHasInstalled", FileMode.Create, FileAccess.Write);
+                fs1.Close();
+            }
+            catch
+            {
+                Process P2 = Process.Start("resources\\python373.exe");
+                P2.WaitForExit();
+                Process P3 = Process.Start("python.exe", "python\\pakcheck.py");
+                P3.WaitForExit();
+            }
+        }
         public void StartEngine()
         {
             try
@@ -42,7 +61,21 @@ namespace ProjectA
                 //启动程序
                 P.Start();
                 //P = Process.Start("python\\pipe.py", "\\\\.\\pipe\\" + PipeName);
-                Client.Connect(timeout*10);
+                //Client.Connect(timeout*10);
+                while (!Client.IsConnected)
+                {
+                    if (!P.HasExited)
+                        try
+                        {
+                            Client.Connect(timeout * 10);
+                        }
+                        catch
+                        {
+                            ;
+                        }
+                    else break;
+                }
+                
                 string s = "csharpPipe";
 
                 result.Initialize();
@@ -87,12 +120,12 @@ namespace ProjectA
         }
         public void CloseEngine()
         {
-            P.Kill();
+            if (!P.HasExited) P.Kill();
             Client.Close();
         }
         ~PipeEngine()
         {
-            P.Kill();
+            if(!P.HasExited) P.Kill();
             Client.Close();
         }
     }
